requirements.txt
# Core dependencies
langchain>=0.1.0
langchain-community>=0.0.13
langchain-core>=0.1.10
langchain-huggingface>=0.0.6
pydantic>=2.5.2

# Vector stores
faiss-cpu>=1.7.4
chromadb>=0.4.22

# Document processing
pymupdf>=1.23.7
pypdf>=3.17.1
langchain-text-splitters>=0.0.1

# Embeddings and LLMs
sentence-transformers>=2.2.2
huggingface-hub>=0.20.2
transformers>=4.36.2
accelerate>=0.25.0
torch>=2.1.2
openai>=1.3.7
cohere>=4.37
langchain_openai==0.3.8
langchain-google-genai==2.1.0

# AWS integration (optional)
boto3>=1.34.1
langchain-aws>=0.0.3

# Utilities
pyyaml>=6.0.1
python-dotenv>=1.0.0
tqdm>=4.66.1